refactor : lint

+++ contracts/interfaces/IStakeTON.sol
@@ -71,6 +71,6 @@ interface IStakeTON {
         uint256 amountOutMinimum,
         uint256 deadline,
         uint160 sqrtPriceLimitX96,
-        uint _type
+        uint256 _type
     ) external returns (uint256 amountOut);
 }
